requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-payments==3.0.1
django-phonenumber-field==8.0.0
django-restframework==0.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.1
phonenumberslite==8.13.45
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
